[Rename] getData API 파일 하나로 통합
- headLine, journal 데이터를 받아오던 api js 파일을 하나로 통일
