--- src/api/getJournalListData.js
@@ -1,11 +1,0 @@
-import { Journal } from "../components/journalList/journalClass.js";
-
-export const getJournal = (URL) => {
-  return fetch(URL)
-    .then((response) => response.json())
-    .then((jsonData) => {
-      const journal = new Journal(jsonData);
-      return journal.makeJournal();
-    })
-    .catch((error) => console.error(`fetch 에러! ${error}`));
-};
